diet_api_project/requirements.txt
django==5.0.1
djangorestframework==3.14.0
google-generativeai==0.3.2
python-dotenv==1.0.0